requirements.txt
llama-index
llama-index-llms-gemini
llama-index-embeddings-gemini
streamlit
openai
python-dotenv
google-generativeai
langchain-google-genai
fitz
pymupdf